Move new cli interface into place
